stream: library API use buffer.  Fix for #10.
